Exit launcher script with google-chrome's exit status
Ctrl+C when running google-chrome actually causes it to exit 0
instead of exit 1.

trapping Ctrl+C and sending kill signal to all child process
(jobs -p or pgrep -P $$) is
risky, so for now, we deal with Ctrl+C or SIGINT to google-chrome
as normal exit.

SIGTERM to the launcher, gchrome will kill the script but leave chrome
running.

google-chrome returns a non-zero exit code only when it runs
in internal error and forced to crash (????)